+++ features/__init__.py
@@ -4,6 +4,7 @@ from . import epic_games
 from . import gholametam
 from . import hafez
 from . import hekmat
+from . import help
 from . import klei_points
 from . import management
 from . import member_count
@@ -21,6 +22,7 @@ epic_games.activate()
 gholametam.activate()
 hafez.activate()
 hekmat.activate()
+help.activate()
 klei_points.activate()
 management.activate()
 member_count.activate()
