+++ features/help/feature.py
@@ -1,0 +1,18 @@
+import logging
+
+_logger = logging.getLogger("main")
+
+
+if "_acive" not in dir():  # Run once
+    global _active
+    _active = False
+
+
+def is_active():
+    return _active
+
+
+def activate():
+    global _active
+    _active = True
+    _logger.debug("features: Feature has been activated: 'help'")
